split observable and doneSameDay, cretaedSameDay check
