+++ src/lib/StatsDashboardBlock/StatsDashboardMainBlock.svelte
@@ -1,20 +1,15 @@
 <script lang="ts">
     import { onDestroy, onMount } from "svelte";
-    import { allTasksList, type Task } from "../../stores/persistentTasks";
+    import { allTasksList } from "../../stores/persistentTasks";
     import StatsDayTasksBlock from "./StatsDayTasksBlock.svelte";
     import StatsOverallDayBlock from "./StatsOverallDayBlock.svelte";
     import StatsRangeSelectorBlock from "./StatsRangeSelectorBlock.svelte";
-    import { Observable, Subject, Subscription, combineLatest, from, map } from "rxjs";
-
-    type StatTask = Task & {
-        hasFinishedThatDay:boolean
-        hasBeenCreatedThatDay:boolean
-        hasBeenActiveThatDay:boolean
-    }
+    import { Observable, Subject, Subscription, from } from "rxjs";
+    import { inRangeTasksObservables, type RangeSelection, type StatTask } from "../../utils/statsObservables";
     
     let statsTargetDate:Date;
     let statsTargetCategories:string[];
-    let rangeSelector:Subject<{ targetDate:Date, targetCategories:string[] }> = new Subject();
+    let rangeSelector:Subject<RangeSelection> = new Subject();
     let inRangeTasksObservable:Observable<StatTask[]>;
     let statsSubscription:Subscription;
 
@@ -22,25 +17,7 @@
     $: if(statsTargetDate || statsTargetCategories) rangeSelector.next({ targetDate: statsTargetDate, targetCategories: statsTargetCategories });
 
     onMount(() => {
-        inRangeTasksObservable = combineLatest([from(allTasksList), rangeSelector]).pipe(
-            map(([allTasks, targetRange]) => allTasks.map(task => {
-                // match all named categories or no category (none)
-                const matchCategoryRange:boolean = (task.category !== null) 
-                    ? targetRange.targetCategories.includes(task.category) 
-                    : targetRange.targetCategories.includes('none');
-                // relevant because achieved that day (even if no activity)
-                const hasFinishedThatDay:boolean = false; // TODO
-                // relevant because created that day (even if no activity)
-                const hasBeenCreatedThatDay:boolean = false; // TODO
-                // relevant because active that day
-                const hasBeenActiveThatDay:boolean = true; // TODO
-                const matchDateRange:boolean = (hasFinishedThatDay || hasBeenCreatedThatDay || hasBeenActiveThatDay);
-
-                return (matchCategoryRange && matchDateRange)
-                    ? {...task, hasFinishedThatDay, hasBeenCreatedThatDay, hasBeenActiveThatDay }
-                    : (null as unknown as StatTask);
-            }).filter(relevantTask => (relevantTask !== null)))
-        );
+        inRangeTasksObservable = inRangeTasksObservables(from(allTasksList), rangeSelector);
         statsSubscription = inRangeTasksObservable.subscribe({ next: val => console.log('val update', val) });
     })
